Changes in epoch-based data reading. bed_to_sparse and deps were simply made compatible, mixed representation for BayesR probably not working

+++ src/densities.hpp
@@ -48,16 +48,16 @@ double alpha_dens(double x, void *norm_data)
 double beta_dens(double x, void *norm_data)
 {
 	pars_beta_sparse p = *(static_cast<pars_beta_sparse *>(norm_data));
-	long double s = 0;
-	long double H = expl((long double) (-p.alpha * x / p.sd));
-	long double G = expl((long double) (p.alpha * p.mean * x / p.sd));
+	double s = 0;
+	double H = exp(-p.alpha * x / p.sd);
+	double G = exp(p.alpha * p.mean * x / p.sd);
 
 	if (p.mixture_value_other != 0)
 	{
 		s = p.rho * sqrt(p.sigmaG1 *p.mixture_value / p.mixture_value_other/ p.sigmaG2) * p.beta_other;
 	}
 
-	long double exp_sums =  G * (p.vi_0 + p.vi_tau_0 + H * (p.vi_1 + p.vi_tau_1 + H * (p.vi_2 + p.vi_tau_2)));
+	double exp_sums =  G * (p.vi_0 + p.vi_tau_0 + H * (p.vi_1 + p.vi_tau_1 + H * (p.vi_2 + p.vi_tau_2)));
 	return -p.alpha * x * p.sum_failure - exp_sums - (x - s) * (x - s) / (2 * p.mixture_value * p.sigmaG1 * (1 - p.rho * p.rho));
 };
 
@@ -65,16 +65,16 @@ double beta_dens2(double x, void *norm_data)
 {
 	pars_beta_sparse p = *(static_cast<pars_beta_sparse *>(norm_data));
 	
-	long double s = 0;
-	long double H = expl((long double)(-p.alpha * x / p.sd));
-	long double G = expl((long double)(p.alpha * p.mean * x / p.sd));
+	double s = 0;
+	double H = exp(-p.alpha * x / p.sd);
+	double G = exp(p.alpha * p.mean * x / p.sd);
 
 	if (p.mixture_value_other != 0)
 	{
 		s = p.rho *  sqrt(p.sigmaG2 * p.mixture_value / p.mixture_value_other / p.sigmaG1) * p.beta_other ;
 	}
 
-	long double exp_sums =  G * (p.vi_0 - p.vi_tau_0 + H * (p.vi_1 - p.vi_tau_1 + H * (p.vi_2 - p.vi_tau_2)));
+	double exp_sums =  G * (p.vi_0 - p.vi_tau_0 + H * (p.vi_1 - p.vi_tau_1 + H * (p.vi_2 - p.vi_tau_2)));
 
 	return -p.alpha * x * p.sum_failure - exp_sums - (x - s) * (x - s) / (2 * p.mixture_value * p.sigmaG2 * (1 - p.rho * p.rho));
 };
